Move fastapi/utils.py to utils/fastapi.py
